fix package structure
